requirements.txt
Flask==3.0.0
pypdf==4.0.1
google-generativeai==0.3.2
google-cloud-texttospeech==2.16.3
pyttsx3==2.90
python-dotenv==1.1.1
gunicorn==21.2.0
